+++ frontend/src/components/LoginCharacterArt.css
@@ -1,0 +1,66 @@
+.login-art-stage {
+  position: absolute;
+  left: clamp(50px, 0 vw, 60px);
+  bottom: clamp(-100px, -15vw, -200px);
+  z-index: 0;
+  width: clamp(130px, 22vw, 230px);
+  aspect-ratio: 231 / 228;
+  display: block;
+  overflow: visible;
+  pointer-events: none;
+}
+
+.login-art-stage .art {
+  width: 100%;
+  height: 100%;
+  display: block;
+  overflow: visible;
+  touch-action: none;
+}
+
+.login-art-stage .art > rect {
+  fill: transparent;
+}
+
+.login-art-stage .character,
+.login-art-stage .body-motion,
+.login-art-stage .face-motion,
+.login-art-stage .pupil,
+.login-art-stage .mouth-motion,
+.login-art-stage .cheeks-motion,
+.login-art-stage .smile-eye-follow {
+  transform-box: view-box;
+  will-change: transform;
+}
+
+.login-art-stage .body-motion {
+  transition: filter 180ms ease;
+}
+
+.login-art-stage .eye-white,
+.login-art-stage .eye-dot,
+.login-art-stage .mouth,
+.login-art-stage .smile-eye {
+  transform-box: fill-box;
+  transform-origin: center;
+}
+
+@media (max-width: 560px) {
+  .login-art-stage {
+    right: -8px;
+    bottom: -36px;
+    width: 150px;
+  }
+}
+
+@media (prefers-reduced-motion: reduce) {
+  .login-art-stage .character,
+  .login-art-stage .body-motion,
+  .login-art-stage .face-motion,
+  .login-art-stage .pupil,
+  .login-art-stage .mouth-motion,
+  .login-art-stage .cheeks-motion,
+  .login-art-stage .smile-eye-follow {
+    transition: none;
+  }
+}
